==> scripts/00-install_packages.R <==
#### Preamble ####
# Purpose: Installs the necessary packages. 
# Author: Rahma Binth Mohammad
# Date: 16 February 2024
# Contact: rahma.binthmohammad@mail.utoronto.ca
# License: MIT

#### Download packages ####
install.packages("tidyverse")
install.packages("dplyr")
install.packages("haven")
install.packages("ggplot2")
install.packages("sf")
install.packages("gridExtra")
install.packages("patchwork")
